requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.12
requests==2.31.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0
pdfplumber==0.7.8
beautifulsoup4==4.12.2
langchain==0.0.341
openai==1.9.0
tiktoken==0.4.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pandas==2.2.3